shelve changes to: Arm backend: Add a repro command when VGF model-converter fails (#20443)

+++ backends/nxp/backend/ir/converter/quantization_utils.py
@@ -1,4 +1,4 @@
-# Copyright 2023-2026 NXP
+# Copyright 2023-2025 NXP
 #
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
@@ -135,12 +135,11 @@ def set_quantization_parameters_to_tensor(
 def quantize_int8(
     data: np.ndarray, scale: List[float], zero_point: List[int]
 ) -> np.ndarray:
-    # noinspection PyTypeChecker
     return quantize(data, zero_point=zero_point, scale=scale)
 
 
 def quantize(
-    value: np.ndarray | float,
+    value: np.ndarray | int,
     zero_point: List[int] | int,
     scale: List[float] | float,
     quant_min: int = -128,
